+++ android/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>TKELogGPS</name>
+	<name>TKELogTrackingData</name>
 	<comment>Project android created by Buildship.</comment>
 	<projects>
 	</projects>
