+++ android/app/BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.tks.tkeloggps",
+    package = "com.tke.gps",
 )
 
 android_resource(
     name = "res",
-    package = "com.tks.tkeloggps",
+    package = "com.tke.gps",
     res = "src/main/res",
 )
