+++ android/app/build.gradle
@@ -102,7 +102,7 @@ android {
     }
 
     defaultConfig {
-        applicationId "com.tks.tkeloggps"
+        applicationId "com.tke.gps"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 1
