+++ android/app/google-services.json
@@ -8,9 +8,9 @@
   "client": [
     {
       "client_info": {
-        "mobilesdk_app_id": "1:656674485036:android:c322c4319c0aef5a",
+        "mobilesdk_app_id": "1:656674485036:android:103c11ad17adea36",
         "android_client_info": {
-          "package_name": "com.tkeloggpsapp"
+          "package_name": "com.tke.gps"
         }
       },
       "oauth_client": [
