+++ android/app/src/main/java/com/tke/gps/MainActivity.java
@@ -1,4 +1,4 @@
-package com.tks.tkeloggps;
+package com.tke.gps;
 
 import com.facebook.react.ReactActivity;
 import com.facebook.react.ReactActivityDelegate;
@@ -13,7 +13,7 @@ public class MainActivity extends ReactActivity {
      */
     @Override
     protected String getMainComponentName() {
-        return "TKELogGPS";
+        return "TKELogTrackingData";
     }
 
     @Override
