+++ android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">TKELogGPS</string>
+    <string name="app_name">TKELog Tracking Data</string>
 </resources>
