+++ android/settings.gradle
@@ -1,4 +1,4 @@
-rootProject.name = 'TKELogGPS'
+rootProject.name = 'TKELogTrackingData'
 include ':react-native-vector-icons'
 project(':react-native-vector-icons').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-vector-icons/android')
 include ':react-native-gesture-handler'
