+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "TKELogGPS",
-  "displayName": "TKELogGPS"
+  "name": "TKELogTrackingData",
+  "displayName": "TKELogTrackingData07"
 }
