+++ ios/TKELogTrackingData.xcodeproj/xcshareddata/xcschemes/TKELogTracking-tvOS.xcscheme
@@ -29,9 +29,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-               BuildableName = "TKELogGPS-tvOS.app"
-               BlueprintName = "TKELogGPS-tvOS"
-               ReferencedContainer = "container:TKELogGPS.xcodeproj">
+               BuildableName = "TKELogTracking-tvOS.app"
+               BlueprintName = "TKELogTracking-tvOS"
+               ReferencedContainer = "container:TKELogTracking.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
@@ -43,9 +43,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
-               BuildableName = "TKELogGPS-tvOSTests.xctest"
-               BlueprintName = "TKELogGPS-tvOSTests"
-               ReferencedContainer = "container:TKELogGPS.xcodeproj">
+               BuildableName = "TKELogTracking-tvOSTests.xctest"
+               BlueprintName = "TKELogTracking-tvOSTests"
+               ReferencedContainer = "container:TKELogTracking.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -61,9 +61,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
-               BuildableName = "TKELogGPS-tvOSTests.xctest"
-               BlueprintName = "TKELogGPS-tvOSTests"
-               ReferencedContainer = "container:TKELogGPS.xcodeproj">
+               BuildableName = "TKELogTracking-tvOSTests.xctest"
+               BlueprintName = "TKELogTracking-tvOSTests"
+               ReferencedContainer = "container:TKELogTracking.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -71,9 +71,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "TKELogGPS-tvOS.app"
-            BlueprintName = "TKELogGPS-tvOS"
-            ReferencedContainer = "container:TKELogGPS.xcodeproj">
+            BuildableName = "TKELogTracking-tvOS.app"
+            BlueprintName = "TKELogTracking-tvOS"
+            ReferencedContainer = "container:TKELogTracking.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <AdditionalOptions>
@@ -94,9 +94,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "TKELogGPS-tvOS.app"
-            BlueprintName = "TKELogGPS-tvOS"
-            ReferencedContainer = "container:TKELogGPS.xcodeproj">
+            BuildableName = "TKELogTracking-tvOS.app"
+            BlueprintName = "TKELogTracking-tvOS"
+            ReferencedContainer = "container:TKELogTracking.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
       <AdditionalOptions>
@@ -113,9 +113,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "TKELogGPS-tvOS.app"
-            BlueprintName = "TKELogGPS-tvOS"
-            ReferencedContainer = "container:TKELogGPS.xcodeproj">
+            BuildableName = "TKELogTracking-tvOS.app"
+            BlueprintName = "TKELogTracking-tvOS"
+            ReferencedContainer = "container:TKELogTracking.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
