+++ ios/TKELogTrackingData.xcodeproj/xcshareddata/xcschemes/TKELogTracking.xcscheme
@@ -29,9 +29,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "TKELogGPS.app"
-               BlueprintName = "TKELogGPS"
-               ReferencedContainer = "container:TKELogGPS.xcodeproj">
+               BuildableName = "TKELogTracking.app"
+               BlueprintName = "TKELogTracking"
+               ReferencedContainer = "container:TKELogTracking.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
@@ -43,9 +43,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "TKELogGPSTests.xctest"
-               BlueprintName = "TKELogGPSTests"
-               ReferencedContainer = "container:TKELogGPS.xcodeproj">
+               BuildableName = "TKELogTrackingTests.xctest"
+               BlueprintName = "TKELogTrackingTests"
+               ReferencedContainer = "container:TKELogTracking.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -61,9 +61,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "TKELogGPSTests.xctest"
-               BlueprintName = "TKELogGPSTests"
-               ReferencedContainer = "container:TKELogGPS.xcodeproj">
+               BuildableName = "TKELogTrackingTests.xctest"
+               BlueprintName = "TKELogTrackingTests"
+               ReferencedContainer = "container:TKELogTracking.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -71,9 +71,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "TKELogGPS.app"
-            BlueprintName = "TKELogGPS"
-            ReferencedContainer = "container:TKELogGPS.xcodeproj">
+            BuildableName = "TKELogTracking.app"
+            BlueprintName = "TKELogTracking"
+            ReferencedContainer = "container:TKELogTracking.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <AdditionalOptions>
@@ -94,9 +94,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "TKELogGPS.app"
-            BlueprintName = "TKELogGPS"
-            ReferencedContainer = "container:TKELogGPS.xcodeproj">
+            BuildableName = "TKELogTracking.app"
+            BlueprintName = "TKELogTracking"
+            ReferencedContainer = "container:TKELogTracking.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
       <AdditionalOptions>
@@ -113,9 +113,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "TKELogGPS.app"
-            BlueprintName = "TKELogGPS"
-            ReferencedContainer = "container:TKELogGPS.xcodeproj">
+            BuildableName = "TKELogTracking.app"
+            BlueprintName = "TKELogTracking"
+            ReferencedContainer = "container:TKELogTracking.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
