+++ ios/TKELogTrackingData/Images.xcassets/AppIcon.appiconset/Contents.json
