+++ ios/TKELogTrackingData/Images.xcassets/Contents.json
