+++ ios/TKELogTrackingDataTests/Info.plist
