+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "TKELogGPS",
+  "name": "TKELogTrackingData",
   "version": "0.0.1",
   "private": true,
   "scripts": {
